backend/requirements.txt
fastapi
uvicorn
faiss-cpu
pypdf
sentence-transformers
python-multipart
llama-cpp-python
